agents/requirements.txt
fastapi
uvicorn
kafka-python
requests
pydantic
python-dotenv
google-generativeai
google-adk
